+++ src/components/GroupsTable/GroupsTable.js
@@ -18,12 +18,21 @@ import {
     PrimaryToolbar
 } from '@redhat-cloud-services/frontend-components';
 import debounce from 'lodash/debounce';
+import difference from 'lodash/difference';
+import flatten from 'lodash/flatten';
+import map from 'lodash/map';
+import union from 'lodash/union';
 import upperCase from 'lodash/upperCase';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { TABLE_DEFAULT_PAGINATION } from '../../constants';
 import { fetchGroups } from '../../store/inventory-actions';
+import useFetchBatched from '../../Utilities/hooks/useFetchBatched';
+import CreateGroupModal from '../InventoryGroups/Modals/CreateGroupModal';
+import DeleteGroupModal from '../InventoryGroups/Modals/DeleteGroupModal';
+import RenameGroupModal from '../InventoryGroups/Modals/RenameGroupModal';
+import { getGroups } from '../InventoryGroups/utils/api';
 import { generateLoadingRows } from '../InventoryTable/helpers';
 import NoEntitiesFound from '../InventoryTable/NoEntitiesFound';
 
@@ -48,9 +57,10 @@ const GROUPS_TABLE_COLUMNS = [
 ];
 
 const GROUPS_TABLE_COLUMNS_TO_URL = {
-    0: 'name',
-    1: 'host_ids',
-    2: 'updated_at'
+    0: '', // reserved for selection boxes
+    1: 'name',
+    2: 'host_ids',
+    3: 'updated_at'
 };
 
 const REQUEST_DEBOUNCE_TIMEOUT = 500;
@@ -62,7 +72,13 @@ const GroupsTable = () => {
     );
     const [filters, setFilters] = useState(GROUPS_TABLE_INITIAL_STATE);
     const [rows, setRows] = useState([]);
+    const [selectedIds, setSelectedIds] = useState([]);
+    const [selectedGroup, setSelectedGroup] = useState({});
+    const [createModalOpen, setCreateModalOpen] = useState(false);
+    const [renameModalOpen, setRenameModalOpen] = useState(false);
+    const [deleteModalOpen, setDeleteModalOpen] = useState(false);
     const groups = useMemo(() => data?.results || [], [data]);
+    const { fetchBatched } = useFetchBatched();
 
     const fetchData = useCallback(
         debounce((filters) => {
@@ -94,10 +110,18 @@ const GroupsTable = () => {
                 </span>,
                 <span key={index}>{(group.host_ids || []).length.toString()}</span>,
                 <span key={index}>{<DateFormat date={group.updated_at} />}</span>
-            ]
+            ],
+            groupId: group.id,
+            groupName: group.name,
+            selected: selectedIds.includes(group.id)
         }));
         setRows(newRows);
-    }, [groups]);
+
+        setSelectedGroup({
+            id: selectedIds[0],
+            name: groups.find(({ id }) => id === selectedIds[0])?.name
+        });
+    }, [groups, selectedIds]);
 
     // TODO: convert initial URL params to filters
 
@@ -172,7 +196,7 @@ const GroupsTable = () => {
                             cells: [
                                 {
                                     title: rejected ? (
-                                        // TODO: don't render the primary button (requires change in FF)
+                                    // TODO: don't render the primary button (requires change in FF)
                                         <ErrorState />
                                     ) : (
                                         <NoEntitiesFound
@@ -193,8 +217,54 @@ const GroupsTable = () => {
 
     // TODO: use ouiaSafe to indicate the loading state for e2e tests
 
+    const onSelect = (event, isSelected, rowId, rowData) => {
+        const { groupId } = rowData;
+        if (isSelected) {
+            setSelectedIds(union(selectedIds, [groupId]));
+        } else {
+            setSelectedIds(difference(selectedIds, [groupId]));
+        }
+    };
+
+    const fetchAllGroupIds = useCallback((filters, total) => {
+        const { sortIndex, sortDirection, perPage, page, ...search } = filters;
+        // exclude sort parameters
+
+        return fetchBatched(getGroups, total, search);
+    }, []);
+
+    const selectAllIds = async () => {
+        const results = await fetchAllGroupIds(filters, data?.total);
+        const ids = map(flatten(map(results, 'results')), 'id');
+        setSelectedIds(ids);
+    };
+
+    const allSelected = selectedIds.length === data?.total;
+    const noneSelected = selectedIds.length === 0;
+    const displayedIds = map(rows, 'groupId');
+    const pageSelected = difference(displayedIds, selectedIds).length === 0;
+
     return (
         <div id="groups-table">
+            <CreateGroupModal
+                isModalOpen={createModalOpen}
+                setIsModalOpen={setCreateModalOpen}
+                reloadData={() => {fetchData(filters);}}
+            />
+            <RenameGroupModal
+                isModalOpen={renameModalOpen}
+                setIsModalOpen={setRenameModalOpen}
+                reloadData={() => fetchData(filters)}
+                modalState={selectedGroup}
+            />
+            <DeleteGroupModal
+                isModalOpen={deleteModalOpen}
+                setIsModalOpen={setDeleteModalOpen}
+                reloadData={() => fetchData(filters)}
+                modalState={selectedIds.length > 1 ? {
+                    ids: selectedIds
+                } : selectedGroup}
+            />
             <PrimaryToolbar
                 pagination={{
                     itemCount: data?.total || 0,
@@ -208,6 +278,68 @@ const GroupsTable = () => {
                 }}
                 filterConfig={{ items: filterConfigItems }}
                 activeFiltersConfig={activeFiltersConfig}
+                bulkSelect={{
+                    items: [
+                        {
+                            title: 'Select none',
+                            onClick: () => setSelectedIds([]),
+                            props: { isDisabled: noneSelected }
+                        },
+                        {
+                            title: `${pageSelected ? 'Deselect' : 'Select'} page (${data?.count || 0} items)`,
+                            onClick: () => {
+                                if (pageSelected) {
+                                    // exclude groups on the page from the selected ids
+                                    const newRows = difference(selectedIds, displayedIds);
+                                    setSelectedIds(newRows);
+                                } else {
+                                    setSelectedIds(union(selectedIds, displayedIds));
+                                }
+                            }
+                        },
+                        {
+                            title: `${allSelected ? 'Deselect' : 'Select'} all (${data?.total || 0} items)`,
+                            onClick: async () => {
+                                if (allSelected) {
+                                    setSelectedIds([]);
+                                } else {
+                                    await selectAllIds();
+                                }
+                            }
+                        }
+                    ],
+                    checked: selectedIds.length > 0, // TODO: support partial selection (dash sign) in FEC BulkSelect
+                    onSelect: async (checked) => {
+                        if (checked) {
+                            await selectAllIds();
+                        } else {
+                            setSelectedIds([]);
+                        }
+                    },
+                    ouiaId: 'groups-selector',
+                    count: selectedIds.length
+                }}
+                actionsConfig={{
+                    actions: [
+                        {
+                            label: 'Create group',
+                            onClick: () => setCreateModalOpen(true)
+                        },
+                        {
+                            label: 'Rename group',
+                            onClick: () => setRenameModalOpen(true),
+                            props: {
+                                isDisabled: selectedIds.length !== 1
+                            }
+                        },
+                        {
+                            label: selectedIds.length > 1 ? 'Delete groups' : 'Delete group',
+                            onClick: () => setDeleteModalOpen(true),
+                            props: {
+                                isDisabled: selectedIds.length === 0
+                            }
+                        }
+                    ] }}
             />
             <Table
                 aria-label="Groups table"
@@ -222,6 +354,30 @@ const GroupsTable = () => {
                 }}
                 onSort={onSort}
                 isStickyHeader
+                onSelect={onSelect}
+                canSelectAll={false}
+                actions={[
+                    {
+                        title: 'Rename group',
+                        onClick: (event, rowIndex, { groupId, groupName }) => {
+                            setSelectedGroup({
+                                id: groupId,
+                                name: groupName
+                            });
+                            setRenameModalOpen(true);
+                        }
+                    },
+                    {
+                        title: 'Delete group',
+                        onClick: (event, rowIndex, { groupId, groupName }) => {
+                            setSelectedGroup({
+                                id: groupId,
+                                name: groupName
+                            });
+                            setDeleteModalOpen(true);
+                        }
+                    }
+                ]}
             >
                 <TableHeader />
                 <TableBody />